Fixed spatial weighting at frame borders
Added SchroOpenGLSpatialWeightPool to pre-create and reuse different spatial
weight blocks and grids

Added schro_opengl_squeeze_pools to free unused items from the pools. This is
currently only necessary for the opengl testsuite because it uses a broad
variety of canvas sizes and spatial block sizes and fills the pools fast with
items that won't be used anymore. If not freed this unused items in the pools
lead to an overallocation of X resources followed by a BadAlloc error

Renamed SchroOpenGLResources to SchroOpenGLCanvasPool

Made blockparsing in shadercompiler more robust