[encoder] Added 1/8th pel MV refinement
Added subpel refinement to 1/8th pel.
Introduced new SchroMe data structure to support
new subpel motion estimation.

Deallocate resources in schro_encoder_frame_complete.
This function is invoked whith the mutex locked; it seems
to improve reliability in multi-threaded runs.
Also small changes to ME to allow considerations of MV
locating reference blocks completely outside of the frame
(this is legal but the encoder only uses a limited extension
around the edges of a frame for ME so the MVs are cropped
when they point outside the frame).